Rename 🚚: header,footer 폴더안 컴포넌트들 이미지 경로 수정
